+++ CHANGELOG.md
@@ -2,3 +2,4 @@ v1.0.2
 ------
 ### Bug fixes
 - Improved URL protocol stripping to better generate clean anchor tag content text.
+- Fix React warnings about missing "key" props for children elements.